+++ Player/Player.h
@@ -5,20 +5,20 @@
 
 class Player {
 private:
-    size_t globalId_;
-    size_t roomId_;
+    int globalId_;
+    int roomId_;
     bool alive_;
     bool isSleep_;
 protected:
     int role_;
 public:
     Player();
-    void setGlobalId(size_t newId);
-    void setRoomId(size_t newId);
+    void setGlobalId(int newId);
+    void setRoomId(int newId);
     void setAlive(bool newStatus);
     void setIsSleep(bool newStatus);
-    size_t getGlobalId();
-    size_t getRoomId();
+    int getGlobalId();
+    int getRoomId();
     bool getAlive();
     bool getIsSleep();
     int getRole();
